requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
anywidget==0.9.18
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autoviz==0.1.720
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
bokeh==2.4.3
cachetools==5.5.2
category-encoders==2.6.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.1.1
dacite==1.9.2
dash==3.0.4
dateutils==0.6.12
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dotenv==0.9.9
duckdb==0.8.1
emoji==2.14.1
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.58.1
fqdn==1.5.1
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.1
grpcio-status==1.71.0
gw_dsl_parser==0.1.6a6
h11==0.16.0
holoviews==1.14.9
htmlmin==0.1.12
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hvplot==0.7.3
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.12.4
importlib-metadata==6.11.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipylab==1.0.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaleido==0.2.1
kanaries_track==0.0.5
kiwisolver==1.4.8
lightgbm==4.6.0
linkify-it-py==2.0.3
llvmlite==0.44.0
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.6.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
monotonic==1.6
more-itertools==10.7.0
multimethod==1.12
narwhals==1.41.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numba==0.61.0
numpy==1.24.4
orjson==3.10.18
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandas-dq==1.28
pandocfilters==1.5.1
panel==0.14.4
param==1.13.0
parso==0.8.4
patsy==1.0.1
phik==0.12.4
pillow==10.4.0
platformdirs==4.3.8
plotly==5.16.1
plotly-resampler==0.10.0
pmdarima==2.0.4
premailer==3.10.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psygnal==0.13.0
pure_eval==0.2.3
puremagic==1.29
pyamg==5.2.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycaret==3.2.0
pycparser==2.22
pyct==0.5.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pygwalker==0.3.11
pyod==2.0.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.2
pyviz_comms==3.0.4
PyWavelets==1.8.0
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
quickjs==1.19.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
schemdraw==0.15
scikit-base==0.5.2
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.9.3
seaborn==0.12.2
segment-analytics-python==2.2.3
Send2Trash==1.8.3
six==1.17.0
sklearn-compat==0.1.3
sktime==0.21.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlglot==26.24.0
stack-data==0.6.3
statsmodels==0.13.5
streamlit==1.45.1
streamlit-option-menu==0.4.0
tbats==1.1.3
tenacity==8.5.0
terminado==0.18.1
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typeguard==4.4.3
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
validators==0.35.0
visions==0.8.1
wasmtime==12.0.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.2
xgboost==1.7.5
xlrd==2.0.1
xxhash==3.5.0
xyzservices==2025.4.0
yagmail==0.15.293
ydata-profiling==4.16.1
yellowbrick==1.5
zipp==3.22.0
